requirements.txt
streamlit
numpy
pandas
torch
torchvision
ftfy
regex
tqdm
Pillow
git+https://github.com/openai/CLIP.git